+++ ouroboros-network/CHANGELOG.md
@@ -12,6 +12,9 @@
 
 * Support for decoding Handshake Query Reply in wireshark dissector.
 * Support for decoding CBOR payload in wireshark dissector.
+* Limit concurrency used by dns resolution.  We only resolve up to 8 dns names
+  concurrently for public / ledger peers and up to 2 for local root peers.
+  This will affect how quickly node connects to ledger peers when it starts.
 
 ## 0.8.1.1